requirements.txt
pynput
requests
python-dotenv